Issue #111 Refactor RawObject-StoredObject relationship to follow Rust norms
